# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music School API",
    "description": "REST API for managing music school students and lessons with GitHub OAuth authentication.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://cse341-w03-music-school-api.onrender.com"
    }
  ],
...
  "paths": {
...
    "/auth/github/callback": {
      "get": {
        "summary": "GitHub OAuth callback",
        "description": "GitHub redirects the user to this endpoint after authentication.",
        "tags": ["Authentication"],
        "responses": {
          "302": {
            "description": "Redirects to the authenticated user's profile."
          },
          "401": {
            "description": "GitHub authentication failed."
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "summary": "Get authenticated user profile",
        "description": "Returns the profile of the currently authenticated GitHub user.",
        "tags": ["Authentication"],
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated user profile returned successfully."
          },
          "401": {
            "description": "User is not authenticated."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "summary": "Log out",
        "description": "Logs out the currently authenticated user and destroys the session.",
        "tags": ["Authentication"],
        "responses": {
          "200": {
            "description": "Successfully logged out."
          },
          "500": {
            "description": "Unable to log out."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Authentication is handled through the Express session after GitHub OAuth login."
      }
    },
...
  }
}